+++ src/bitmap/bitmap.cmake
@@ -15,6 +15,7 @@ if(${INVADER_BITMAP})
         src/bitmap/bitmap.cpp
         src/bitmap/stb/stb_impl.c
         src/bitmap/color_plate_scanner.cpp
+        src/bitmap/sprite.cpp
         src/bitmap/image_loader.cpp
         src/bitmap/bitmap_data_writer.cpp
     )